Revert "Feature/dropbox login fix"
