+++ ImgurUploader/View/ContentView.swift
@@ -7,11 +7,11 @@
 
 import SwiftUI
 import PhotosUI
-import SwiftyDropbox
 
 struct ContentView: View {
     @State private var showingAlert:Bool = false
     @State private var pasteString:String  = ""
+    @State private var showingToolbar:Bool = true
     @State var selectedItem: PhotosPickerItem?
     @State var image: UIImage?
     @State var isSelected: Bool = false
@@ -21,10 +21,9 @@ struct ContentView: View {
     @Environment(\.modelContext) private var modelContext
     @State private var showAd: Bool = false
     @Environment(\.colorScheme) var colorScheme
-    @State private var alertMessage: String = ""
-    @State private var isAlertShowing: Bool = false
     
     var body: some View {
+        
         NavigationStack {
             VStack {
                 Spacer()
@@ -61,6 +60,7 @@ struct ContentView: View {
                             .font(.title)
                         }
                         .onChange(of: selectedItem) {
+                            showingToolbar = false
                             Task {
                                 guard let imageData = try await selectedItem?.loadTransferable(type: Data.self) else { return }
                                 guard let uiImage = UIImage(data: imageData) else { return }
@@ -72,14 +72,16 @@ struct ContentView: View {
                         
                         Button(action: {
                             
-                            // 認証前
-                            if (DropboxClientsManager.authorizedClient == nil) {
-                                dropboxViewModel.performLogin()
-                            } else {
-                                Task {
-                                    dropboxViewModel.listFiles()
-                                }
+                            print(dropboxViewModel.isAuthenticated)
+                            
+                            if dropboxViewModel.isAuthenticated {
+                                // 認証済みなら画像の一覧を取得
+                                dropboxViewModel.listFiles()
                                 isShowDropboxList = true
+                                
+                            } else {
+                                // 未認証なら認証画面を表示
+                                dropboxViewModel.performLogin()
                             }
                             
                         }, label: {
@@ -112,6 +114,7 @@ struct ContentView: View {
                         Button("Cancel") {
                             image = nil
                             isSelected.toggle()
+                            showingToolbar.toggle()
                         }
                         .padding(5)
                         .background(.red)
@@ -126,32 +129,10 @@ struct ContentView: View {
                 
                 BannerAd()
             }
-            .onAppear() {
-            }
-            .interstitialAd(isPresented: $showAd)
-            .onOpenURL { url in
-                print("url: \(url)")
-                let oauthCompletion: DropboxOAuthCompletion = {
-                    if let authResult = $0 {
-                        switch authResult {
-                        case .success:
-                            alertMessage = "Successfully logged in to dropbox."
-                            isAlertShowing = true
-                        case .cancel:
-                            alertMessage = "Authentication to dropbox has been canceled."
-                            isAlertShowing = true
-                        case .error(_, _):
-                            alertMessage = "An unexpected error has occurred."
-                            isAlertShowing = true
-                        }
-                    }
-                }
-                DropboxClientsManager.handleRedirectURL(url, backgroundSessionIdentifier: "patata", completion: oauthCompletion)
-            }
             .toolbar {
                 ToolbarItem(placement: .navigationBarTrailing) {
                     NavigationLink(destination: ListView()) {
-                        if !isSelected {
+                        if showingToolbar {
                             Image(systemName: "photo.stack.fill")
                         }
                     }
@@ -160,12 +141,14 @@ struct ContentView: View {
             .toolbar {
                 ToolbarItem(placement: .navigationBarLeading) {
                     NavigationLink(destination: InfoView()) {
-                        if !isSelected {
+                        if showingToolbar {
                             Image(systemName: "info.circle")
                         }
                     }
                 }
             }
+            .interstitialAd(isPresented: $showAd)
+            
             .sheet(isPresented: $isShowDropboxList) {
                 let files = dropboxViewModel.files
                 let images = dropboxViewModel.dropboxImages
@@ -211,11 +194,15 @@ struct ContentView: View {
                     }
                 }
             }
+
             .sheet(isPresented: $viewModel.isShowSheet,onDismiss: {
                 image = nil
+                showingToolbar = true
                 viewModel.isShowSheet = false
                 isSelected = false
                 showAd = true
+                
+                /// データ永続化
                 let newData = ImageData(url: viewModel.postedImageData!.data.link, deletehas: viewModel.postedImageData!.data.deletehash)
                 modelContext.insert(newData)
             }){
@@ -243,12 +230,6 @@ struct ContentView: View {
                     }
                 }
                 
-            }
-            .alert("Authentication Result.", isPresented: $isAlertShowing) {
-                Button("OK") {}
-                
-            } message: {
-                Text(alertMessage)
             }
         }
     }
