+++ ImgurUploader/View/InfoView.swift
@@ -9,9 +9,8 @@ import SwiftUI
 import WebUI
 
 struct InfoView: View {
-    @StateObject var viewModel = DropboxViewModel()
-    
     var body: some View {
+        
         List {
             NavigationLink(
                 destination: TutorialView(),
@@ -45,15 +44,6 @@ struct InfoView: View {
                     }
                 }
             )
-            
-            Button("Logout Dropbox") {
-                viewModel.logout()
-            }
-        }
-        .alert("Logout succeeded.", isPresented: $viewModel.isLoggedOut) {
-            Button("OK") {
-                
-            }
         }
     }
 }
