+++ ImgurUploader/ViewModel/DropboxViewModel.swift
@@ -14,7 +14,6 @@ class DropboxViewModel: ObservableObject {
     @Published var selectedImage: UIImage? = nil
     @Published var isLoading = false
     @Published var dropboxImages: [String: UIImage] = [:]
-    @Published var isLoggedOut = false
     
     init() {
         checkAuthentication()
@@ -38,9 +37,28 @@ class DropboxViewModel: ObservableObject {
             UIApplication.shared,
             controller: nil,
             loadingStatusDelegate: nil,
-            openURL: { (url: URL) -> Void in UIApplication.shared.open(url, options: [:], completionHandler: nil) },
+            openURL: { (url: URL) -> Void in
+                UIApplication.shared.open(url, options: [:], completionHandler: nil)
+            },
             scopeRequest: scopeRequest
         )
+    }
+    
+    func handleOpenURL(_ url: URL) {
+        let oauthCompletion: DropboxOAuthCompletion = {
+            if let authResult = $0 {
+                switch authResult {
+                case .success:
+                    print("Success! User is logged into DropboxClientsManager.")
+                    self.isAuthenticated = true
+                case .cancel:
+                    print("Authorization flow was manually canceled by user!")
+                case .error(_, let description):
+                    print("Error: \(String(describing: description))")
+                }
+            }
+        }
+        DropboxClientsManager.handleRedirectURL(url, backgroundSessionIdentifier: "patata", completion: oauthCompletion)
     }
     
     func listFiles() {
@@ -62,8 +80,9 @@ class DropboxViewModel: ObservableObject {
                         }
                         return false
                     }
+                    print("Filtered files: \(self.files)")
                     
-                    // ファイルリストが確定した後、各画像を自動的にダウンロード
+                    // ファイルリストを取得した後、各画像を自動的にダウンロード
                     self.files.forEach { file in
                         self.downloadImage(file)
                     }
@@ -73,15 +92,14 @@ class DropboxViewModel: ObservableObject {
             }
         }
     }
-    
-    
+
     
     func downloadImage(_ file: Files.Metadata) {
         guard let client = DropboxClientsManager.authorizedClient else {
             print("User is not logged in")
             return
         }
-        
+
         if let fileMetadata = file as? Files.FileMetadata {
             print("Downloading image for path: \(fileMetadata.pathLower ?? "unknown")")
             client.files.download(path: fileMetadata.pathLower ?? "").response { response, error in
@@ -97,14 +115,5 @@ class DropboxViewModel: ObservableObject {
             }
         }
     }
-    
-    func logout() {
-        DropboxClientsManager.unlinkClients()
-        isAuthenticated = false
-        files.removeAll()
-        dropboxImages.removeAll()
-        isLoggedOut = true
-        print("User logged out from Dropbox.")
-    }
-    
+
 }
